gsd-power: Apply a backlight brightness curve
gsd-power controls display backlight brightness with an API that takes
values from 0-100 and sets the brightness of the backlight linearly
according to that value. This commit introduces a backlight brightness
curve, which allows for the tuning of backlight brightness control. This
allows for improved UX for control by accomodating the non-linearity of
human perception of brightness. In theory, the brightness difference
between each step can be made perceptually identical. This increases
granularity in low brightness levels and responsiveness in high
brightness levels.

The curve is created with the clutter_ease_cubic_bezier() function
inherited from clutter. The curve can be tuned by changing the input
variables of the function. The curve is currently described by the
following parameters bezier(.1, 0.05, 0.87, 0.12).

The brightness curve is applied based on the value of the "scale" sysfs
attribute, and the custom "GSD_BACKLIGHT_SCALE" property of the
backlight. The curve is applied under the following conditions, where
"GSD_BACKLIGHT_SCALE" is given priority to allow user configuration:

+----------------------+--------------------------+
| backlight scale type | brightness curve applied |
+----------------------+--------------------------+
| linear               | yes                      |
| non-linear           | no                       |
| unknown              | yes                      |
+----------------------+--------------------------+