rfkill: take care of airplane mode by calling ofono instead of network manager
now hear me out. i'm not saying the previous method of doing this was bad, not at all. its just that its not very useful on a phone
lets say you are on a plane (which is what that mode is for, surprise surprise) transatlantic flight and your airline has in-flight wifi (as most proper airlines do nowadays)
if we call network manager it will also block wifi which means we are missing out on that juicy airplane wifi
by calling ofono directly we will disable the modem (which is what modern android does too) but keep the rest intact and let the user decide what they want to do with those
previously we took care of this in settings but that was very hacky and very ugly, nor did it work very well

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>