rfkill: refresh wwan_interesting manually on HasAirplaneMode
right now interface-added is not emitted properly. might be an ofono2mm issue (but it seems to be emitting fine?), might be dbus next/fast, i am not sure. but wwan_interesting not updating properly was causing all sorts of odd behaviours
refreshing it manually is a good compromise for now until we figure out how to properly fix this

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>