power: partially revert brightness curve implementation
the assumption here AGAIN is that brightness is a value between 0 and 100 but this time taken to the extreme.
it hardcodes 0 and 100 in min and max, making it impossible to override. this commit as a whole is very questionable.
revert everything we don't need and only keep whatever is needed for linear auto brightness to work

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>